refactor: unified to ATraverser/AReverseTraverser
